+++ typesafety/assert_type/test_msgspec.py
@@ -12,7 +12,7 @@ class _HeaderModel(msgspec.Struct):
 
 
 # Test that type args work:
-_HeaderModel(token=1)  # type: ignore[arg-type]
+_HeaderModel(token=1)  # type: ignore[arg-type]  # ty: ignore[invalid-argument-type]
 
 
 class _QueryModel(msgspec.Struct):