+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Later on we will make the API more stable and decrease the amount
 of requirements for an API to count as public.
 
 
-## WIP
+## 0.15.0 WIP
 
 ### Features